ci(refactor): unique artifact names + LTS/Legacy matrix; compat hub
