+++ report_stats.py
@@ -130,7 +130,7 @@ def build_ozet_df(
             "satis_tarihi",
         ]
     ]
-    from utils.pandas_safe import safe_infer_objects
+    from utils.compat import safe_infer_objects
 
     with option_context("future.no_silent_downcasting", True):
         subset = safe_infer_objects(subset.fillna({"hisse_sayisi": 0}), copy=False)
